Extract input form Fixes #4
